feat(bloomable): Remember sessions to prevent re-logging in. refactor(bloomable): Moved functions to more declarative files.
